34774: LA port of 8337987: Relocate jfr and throw_exception stubs from StubGenerator to SharedRuntime
