+++ src/hotspot/cpu/loongarch/templateInterpreterGenerator_loongarch.cpp
@@ -835,8 +835,8 @@ void TemplateInterpreterGenerator::generate_stack_overflow_check(void) {
   // Note: the restored frame is not necessarily interpreted.
   // Use the shared runtime version of the StackOverflowError.
   __ move(SP, Rsender);
-  assert(StubRoutines::throw_StackOverflowError_entry() != nullptr, "stub not yet generated");
-  __ jmp(StubRoutines::throw_StackOverflowError_entry(), relocInfo::runtime_call_type);
+  assert(SharedRuntime::throw_StackOverflowError_entry() != nullptr, "stub not yet generated");
+  __ jmp(SharedRuntime::throw_StackOverflowError_entry(), relocInfo::runtime_call_type);
 
   // all done with frame size check
   __ bind(after_frame_check);
